requirements.txt
asyncpg>=0.30.0
chainlit>=2.7.1.1
fastapi>=0.115.0
google-genai>=1.53.0
httpx>=0.28.0
numpy>=2.3.3
openai>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
qdrant-client>=1.12.0
uvicorn>=0.30.0
